+++ src/meson.build
@@ -34,6 +34,7 @@ sources = files(
   'pkg/utils/arch.go',
   'pkg/utils/errors.go',
   'pkg/utils/fedora.go',
+  'pkg/utils/rhel.go',
   'pkg/utils/utils.go',
   'pkg/utils/utils_cgo.go',
   'pkg/utils/utils_test.go',
